Updated final tag on profile sections
