+++ src/Componentes/PerfilPremium/ListaCostos.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { View } from 'react-native';
-import * as Componentes from './Indice';
+import * as Componentes from '../Indice';
 
 export default ListaCostos = ({costos}) => {
     return(
         <View>
             {
                 costos.map((c, i) => (
-                    <Componentes.Costo
+                    <Componentes.PerfilPremium.Costo
                         titulo={c.titulo}
                         icono={c.icono}
                         descripcion={c.descripcion}
@@ -15,6 +15,7 @@ export default ListaCostos = ({costos}) => {
                         />
                 ))
             }
+            <Componentes.PerfilTrabajador.FinSeccion />
         </View>
     );
 }
