+++ src/Componentes/PerfilPremium/ListaMultimedia.js
@@ -1,11 +1,8 @@
 import React from 'react';
 import {
-    StyleSheet,
-    Text,
     View,
 } from 'react-native';
 import * as Componentes from '../Indice';
-import Colores from '../../Estilos/Colores';
 
 export default ListaMultimedia = ({ multimedia }) => {
     return(
@@ -17,17 +14,7 @@ export default ListaMultimedia = ({ multimedia }) => {
                         />
                 ))
             }
-            <Text style={Estilos.EtiquetaFinal}>No hay más multimedia</Text>
+            <Componentes.PerfilTrabajador.FinSeccion />
         </View>
     );
 }
-
-const Estilos = StyleSheet.create({
-    EtiquetaFinal: {
-        color: Colores.etiquetas,
-        textAlign: 'center',
-        marginVertical: 10,
-        marginHorizontal: 50,
-        padding: 10,
-    }
-})
