+++ src/Componentes/PerfilPremium/ListaPublicacion.js
@@ -1,15 +1,9 @@
 import React from 'react';
 import {
-    StyleSheet,
     View,
 } from 'react-native';
-import {
-    Text,
-} from 'react-native-elements';
-//////
 import * as Componentes from '../Indice';
-import Colores from '../../Estilos/Colores';
-//////
+
 export default ListaPublicacion = ({avatar,
                                     nombre,
                                     valoracion,
@@ -26,19 +20,7 @@ export default ListaPublicacion = ({avatar,
                         />
                 ))
             }
-            <Text style={Estilos.EtiquetaFinal}>No hay más publicaciones</Text>
+            <Componentes.PerfilTrabajador.FinSeccion />
         </View>
     );
 }
-// ESTILOS
-const Estilos = StyleSheet.create({
-    EtiquetaFinal: {
-        color: Colores.etiquetas,
-        borderTopColor: Colores.separador,
-    /*    borderTopWidth: 2,*/
-        textAlign: 'center',
-        marginVertical: 10,
-        marginHorizontal: 50,
-        padding: 10,
-    }
-});
